reset images on going back to browser to fix some bugs where coming back to the same collection and selecting one of the images doesnt load them cos the information is stale.
reset images on going back to browser to fix some bugs where coming back to the same collection and selecting one of the images doesnt load them cos the information is stale.

reset images on going back to browser to fix some bugs where coming back to the same collection and selecting one of the images doesnt load them cos the information is stale.

reset images on going back to browser to fix some bugs where coming back to the same collection and selecting one of the images doesnt load them cos the information is stale.